tests.cpp: int now long int

+++ C++/problem003/tests.cpp
@@ -6,16 +6,16 @@
 #include <cmath>
 #include "ep003.h"
 
-bool IsFactor(int target, int factor)
+bool IsFactor(long int target, long int factor)
 {
     return (target % factor == 0) ? true : false;
 }
 
-bool IsPrime (int factor)
+bool IsPrime (long int factor)
 {
     // uses Fermat's Little Theorem
     // cast to int should be safe as 2^int is an int
     double fermat = std::exp2(factor);
 
-    return ((int)fermat % factor == 2) ? true : false;
+    return ((long int)fermat % factor == 2) ? true : false;
 }
